@@ -1,4 +1,7 @@
-jQuery(document).ready(function(){
+var ontNotesTable;
+var ont_columns = { archived: 3, date: 7, subjectSort: 2 };
+
+jQuery(".ontologies.show").ready(function(){
   setupNotesFaceboxSizing();
   bindAddCommentClick();
   bindAddProposalClick();
@@ -6,10 +9,16 @@ jQuery(document).ready(function(){
   bindReplyClick();
   bindReplyCancelClick();
   bindReplySaveClick();
-  // Wire up subscriptions button activity
+
   jQuery("a.subscribe_to_notes").live("click", function(){
     subscribeToNotes(this);
   });
+
+  jQuery("#hide_archived_ont").click(function(){
+    hideOrUnhideArchivedOntNotes();
+  });
+
+  wireOntTable(jQuery("#ontology_notes_list"));
 });
 
 NOTES_PROPOSAL_TYPES = {
@@ -364,4 +373,36 @@ function subscribeToNotes(button) {
   });
 }
 
+function hideOrUnhideArchivedOntNotes() {
+  if (jQuery("#hide_archived_ont:checked").val() !== undefined) {
+    // Checked
+    ontNotesTable.fnFilter('false', ont_columns.archived);
+  } else {
+    // Unchecked
+    ontNotesTable.fnFilter('', ont_columns.archived, true, false);
+  }
+}
 
+function wireOntTable(ontNotesTableNew) {
+  jQuery.data(document.body, "ontology_id", "#{@ontology.acronym}");
+
+  ontNotesTable = ontNotesTableNew;
+  ontNotesTable.dataTable({
+    "iDisplayLength": 50,
+    "sPaginationType": "full_numbers",
+    "aaSorting": [[ont_columns.date, 'desc']],
+    "aoColumns": [
+      { "bVisible": false }, // Delete
+      { "iDataSort": ont_columns.subjectSort }, // Subject link
+      { "bVisible": false }, // Subject for sort
+      { "bVisible": false }, // Archived for filter
+      null, // Author
+      null, // Type
+      null, // Target
+      null // Created
+    ],
+  });
+  // Important! Table is somehow getting set to zero width. Reset here.
+  jQuery(ontNotesTable).css("width", "100%");
+  ontNotesTable.fnFilter('false', ont_columns.archived);
+}
