@@ -9,7 +9,7 @@ Bundler.require(*Rails.groups)
 module BioportalWebUi
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.1
+    config.load_defaults 5.2
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
@@ -18,5 +18,7 @@ module BioportalWebUi
 
     # Serve error pages from the Rails app itself, instead of using static error pages in /public.
     config.exceptions_app = self.routes
+
+    config.settings = config_for(:settings)
   end
 end
